REF: implement test_astype (#33734)
