+++ pandas/tests/indexes/test_numeric.py
@@ -1,5 +1,4 @@
 from datetime import datetime, timedelta
-import re
 
 import numpy as np
 import pytest
@@ -9,7 +8,6 @@ from pandas._libs.tslibs import Timestamp
 import pandas as pd
 from pandas import Float64Index, Index, Int64Index, Series, UInt64Index
 import pandas._testing as tm
-from pandas.api.types import pandas_dtype
 from pandas.tests.indexes.common import Base
 
 
@@ -213,67 +211,6 @@ class TestFloat64Index(Numeric):
             mixed_index, Index([1.5, 2, 3, 4, 5], dtype=object), is_float_index=False
         )
 
-    def test_astype(self, mixed_index, float_index):
-
-        result = float_index.astype(object)
-        assert result.equals(float_index)
-        assert float_index.equals(result)
-        self.check_is_index(result)
-
-        i = mixed_index.copy()
-        i.name = "foo"
-        result = i.astype(object)
-        assert result.equals(i)
-        assert i.equals(result)
-        self.check_is_index(result)
-
-        # GH 12881
-        # a float astype int
-        for dtype in ["int16", "int32", "int64"]:
-            i = Float64Index([0, 1, 2])
-            result = i.astype(dtype)
-            expected = Int64Index([0, 1, 2])
-            tm.assert_index_equal(result, expected)
-
-            i = Float64Index([0, 1.1, 2])
-            result = i.astype(dtype)
-            expected = Int64Index([0, 1, 2])
-            tm.assert_index_equal(result, expected)
-
-        for dtype in ["float32", "float64"]:
-            i = Float64Index([0, 1, 2])
-            result = i.astype(dtype)
-            expected = i
-            tm.assert_index_equal(result, expected)
-
-            i = Float64Index([0, 1.1, 2])
-            result = i.astype(dtype)
-            expected = Index(i.values.astype(dtype))
-            tm.assert_index_equal(result, expected)
-
-        # invalid
-        for dtype in ["M8[ns]", "m8[ns]"]:
-            msg = (
-                f"Cannot convert Float64Index to dtype {pandas_dtype(dtype)}; "
-                f"integer values are required for conversion"
-            )
-            with pytest.raises(TypeError, match=re.escape(msg)):
-                i.astype(dtype)
-
-        # GH 13149
-        for dtype in ["int16", "int32", "int64"]:
-            i = Float64Index([0, 1.1, np.NAN])
-            msg = r"Cannot convert non-finite values \(NA or inf\) to integer"
-            with pytest.raises(ValueError, match=msg):
-                i.astype(dtype)
-
-    def test_cannot_cast_inf_to_int(self):
-        idx = pd.Float64Index([1, 2, np.inf])
-
-        msg = r"Cannot convert non-finite values \(NA or inf\) to integer"
-        with pytest.raises(ValueError, match=msg):
-            idx.astype(int)
-
     def test_type_coercion_fail(self, any_int_dtype):
         # see gh-15832
         msg = "Trying to coerce float values to integers"
@@ -358,13 +295,6 @@ class TestFloat64Index(Numeric):
         tm.assert_numpy_array_equal(i.isin([1.0, np.nan]), np.array([True, True]))
         i = Float64Index([1.0, 2.0])
         tm.assert_numpy_array_equal(i.isin([np.nan]), np.array([False, False]))
-
-    def test_astype_from_object(self):
-        index = Index([1.0, np.nan, 0.2], dtype="object")
-        result = index.astype(float)
-        expected = Float64Index([1.0, np.nan, 0.2])
-        assert result.dtype == expected.dtype
-        tm.assert_index_equal(result, expected)
 
     def test_fillna_float64(self):
         # GH 11343
